Refactored Result tests into separate file
